+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Exit menu item. (Windows)
+- Prevent the texture cache from breaking when an asset can't fit into an empty cache page.
 
 ## [1.0.1] - 2022-08-21